Potential fix for code scanning alert no. 2: Multiplication result converted to larger type

+++ src/hw/nvme.c
@@ -468,12 +468,12 @@ nvme_bounce_xfer(struct nvme_namespace *ns, u64 lba, void *buf, u16 count,
     u16 blocks = count < max_blocks ? count : max_blocks;
 
     if (write)
-        memcpy(nvme_dma_buffer, buf, blocks * ns->block_size);
+        memcpy(nvme_dma_buffer, buf, (size_t)blocks * ns->block_size);
 
     int res = nvme_io_xfer(ns, lba, nvme_dma_buffer, NULL, blocks, write);
 
     if (!write && res >= 0)
-        memcpy(buf, nvme_dma_buffer, res * ns->block_size);
+        memcpy(buf, nvme_dma_buffer, (size_t)res * ns->block_size);
 
     return res;
 }
